.archive/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "Orchestrate coding tasks across cheap LLMs with Claude review"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "gitpython>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
forge = "forge.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["forge*"]
